LLM Explanation/requirements.txt
pandas
numpy
matplotlib
seaborn
google-genai
reportlab